move fractal reconstruction formulas to formulae/
